Speed up flash programming

+++ test/automated_test.py
@@ -28,14 +28,15 @@ from test_util import TestResult, Test, Logger
 
 from basic_test import basic_test
 from speed_test import SpeedTest
-#from flash_test import flash_test
+from flash_test import FlashTest
 
 
 if __name__ == "__main__":
     log_file = "automated_test_result.txt"
 
     # Setup logging
-    os.remove(log_file)
+    if os.path.exists(log_file):
+        os.remove(log_file)
     logger = Logger(log_file)
     sys.stdout = logger
     sys.stderr = logger
@@ -48,8 +49,7 @@ if __name__ == "__main__":
     test = Test("Basic Test", lambda board: basic_test(board, None))
     test_list.append(test)
     test_list.append(SpeedTest())
-    #test = Test("Flash Test", flash_test)
-    #test_list.append(test)
+    test_list.append(FlashTest())
 
     # Put together list of boards to test
     board_list = MbedBoard.getAllConnectedBoards(close = True, blocking = False)
